Move common internal api and model classes to the common package
COSDK-500
