+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/AnalyticsManagerFactory.kt
@@ -19,8 +19,8 @@ import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsTrackRequ
 import com.adyen.checkout.core.analytics.internal.data.remote.DefaultAnalyticsRemoteDataStore
 import com.adyen.checkout.core.analytics.internal.data.remote.DefaultAnalyticsSetupProvider
 import com.adyen.checkout.core.analytics.internal.data.remote.api.AnalyticsService
+import com.adyen.checkout.core.common.internal.api.HttpClientFactory
 import com.adyen.checkout.core.components.data.model.Amount
-import com.adyen.checkout.core.internal.data.api.HttpClientFactory
 import com.adyen.checkout.core.internal.ui.model.ComponentParams
 import java.util.Locale
