+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/model/AnalyticsSetupRequest.kt
@@ -8,15 +8,15 @@
 
 package com.adyen.checkout.core.analytics.internal.data.remote.model
 
+import com.adyen.checkout.core.common.internal.model.JsonUtils
+import com.adyen.checkout.core.common.internal.model.ModelObject
+import com.adyen.checkout.core.common.internal.model.ModelUtils
+import com.adyen.checkout.core.common.internal.model.getIntOrNull
+import com.adyen.checkout.core.common.internal.model.getLongOrNull
+import com.adyen.checkout.core.common.internal.model.getStringOrNull
+import com.adyen.checkout.core.common.internal.model.optStringList
 import com.adyen.checkout.core.components.data.model.Amount
 import com.adyen.checkout.core.exception.ModelSerializationException
-import com.adyen.checkout.core.internal.data.model.JsonUtils
-import com.adyen.checkout.core.internal.data.model.ModelObject
-import com.adyen.checkout.core.internal.data.model.ModelUtils
-import com.adyen.checkout.core.internal.data.model.getIntOrNull
-import com.adyen.checkout.core.internal.data.model.getLongOrNull
-import com.adyen.checkout.core.internal.data.model.getStringOrNull
-import com.adyen.checkout.core.internal.data.model.optStringList
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
