+++ core/src/main/java/com/adyen/checkout/core/components/data/PaymentComponentData.kt
@@ -8,14 +8,14 @@
 
 package com.adyen.checkout.core.components.data
 
+import com.adyen.checkout.core.common.internal.model.ModelObject
+import com.adyen.checkout.core.common.internal.model.ModelUtils.deserializeOpt
+import com.adyen.checkout.core.common.internal.model.ModelUtils.serializeOpt
+import com.adyen.checkout.core.common.internal.model.getBooleanOrNull
+import com.adyen.checkout.core.common.internal.model.getStringOrNull
 import com.adyen.checkout.core.components.data.model.Amount
 import com.adyen.checkout.core.components.paymentmethod.PaymentMethodDetails
 import com.adyen.checkout.core.exception.ModelSerializationException
-import com.adyen.checkout.core.internal.data.model.ModelObject
-import com.adyen.checkout.core.internal.data.model.ModelUtils.deserializeOpt
-import com.adyen.checkout.core.internal.data.model.ModelUtils.serializeOpt
-import com.adyen.checkout.core.internal.data.model.getBooleanOrNull
-import com.adyen.checkout.core.internal.data.model.getStringOrNull
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
