+++ core/src/main/java/com/adyen/checkout/core/internal/SessionsPaymentFacilitatorFactory.kt
@@ -9,9 +9,9 @@
 package com.adyen.checkout.core.internal
 
 import androidx.lifecycle.SavedStateHandle
+import com.adyen.checkout.core.common.internal.api.HttpClientFactory
 import com.adyen.checkout.core.components.CheckoutCallback
 import com.adyen.checkout.core.components.CheckoutConfiguration
-import com.adyen.checkout.core.internal.data.api.HttpClientFactory
 import com.adyen.checkout.core.internal.ui.model.SessionParamsFactory
 import com.adyen.checkout.core.sessions.CheckoutSession
 import com.adyen.checkout.core.sessions.SessionInteractor
