+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/data/api/SessionService.kt
@@ -10,8 +10,8 @@ package com.adyen.checkout.core.sessions.internal.data.api
 
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.DispatcherProvider
-import com.adyen.checkout.core.internal.data.api.HttpClient
-import com.adyen.checkout.core.internal.data.api.post
+import com.adyen.checkout.core.common.internal.api.HttpClient
+import com.adyen.checkout.core.common.internal.api.post
 import com.adyen.checkout.core.sessions.internal.data.model.SessionPaymentsRequest
 import com.adyen.checkout.core.sessions.internal.data.model.SessionPaymentsResponse
 import com.adyen.checkout.core.sessions.internal.data.model.SessionSetupRequest
